backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
pillow==10.3.0
opencv-python==4.9.0.80
werkzeug==3.0.2
oss2==2.18.3
requests==2.31.0
